django_app/requirements.txt
Django==4.2.4
django-multitenant==3.2.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
uWSGI==2.0.22
